requirements.txt
sqlalchemy
pandas
snowflake-snowpark-python[pandas]